requirements.txt
flask==2.3.3
python-dotenv==1.0.0
requests==2.31.0
python-docx==0.8.11
PyPDF2==3.0.1
python-magic==0.4.27
flask-cors==4.0.0
